+++ include/codi/tapes/data/position.hpp
@@ -37,8 +37,8 @@
 #include <cstdlib>
 #include <iostream>
 
-#include "../../misc/macros.hpp"
 #include "../../config.h"
+#include "../../misc/macros.hpp"
 
 /** \copydoc codi::Namespace */
 namespace codi {