+++ include/codi/tapes/misc/duplicateJacobianRemover.hpp
@@ -36,8 +36,8 @@
 
 #include <array>
 
-#include "../../misc/macros.hpp"
 #include "../../config.h"
+#include "../../misc/macros.hpp"
 
 /** \copydoc codi::Namespace */
 namespace codi {